requirements.txt
# Core dependencies
flask==2.3.3
flask-cors==4.0.0

# LLM and RAG
transformers==4.35.2
torch==2.2.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.1
huggingface-hub==0.19.4

# Vector storage and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3